rename sndcore module as audiocontext
